+++ CHANGELOG.md
@@ -7,5 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-[Unreleased]: https://github.com/rokwire/rokwire-sdk-go/compare/v1.0.0....HEAD
-[1.0.0]: https://github.com/rokwire/rokwire-sdk-go/tree/v1.0.0
+## [1.0.0] - 2024-07-30
+### Changed
+- Update existing packages [#1](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/1)
+
+### Added
+- Add common code from BB template [#3](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/3)  
+- Auth library
+- Logging library
+
+[1.0.0]: https://github.com/rokwire/rokwire-building-block-sdk-go/tree/v1.0.0